Extract trigram search code into dedicated module
Move trigram indexing and search functions from filter::tree into a
new filter::trigram module. Update all call sites in josh-filter and
graphql to use the new module path.

Change: trigram-extract
